Code restructured to fix deployment errors
